Add internal QR code reader
WALLETH uses the external application Barcode Scanner to read QR codes.
This dependency creates an awkward UX if the user does not already have
it installed on their device. This patch implements a basic QR code
scanner activity to be used for application-wide scanning and removes
the dependency on Barcode Scanner.

The scanner does require the camera permission, which is in the
[dangerous permission] group. This can be granted or denied at install
time on devices running API level < 23 or at run time on devices running
API level >= 23.

Some future improvements to consider:

  * Determine an ideal camera preview size. The default size is provided
  by the camera driver, which is most commonly 1920x1080 and may or may
  not match the size of the preview surface.

  * Maintain the aspect ratio of the preview surface. The surface
  stretches to the size of its parent container. On most devices, the
  distortion is minimal since the preview size and display aspect ratios
  are very similar or the same.

  * Handle 180° device rotations. The camera display orientation is not
  updated unless the activity is restarted, which is not triggered when
  rotating the device 180°. This means the camera preview will appear
  upside down, though scanning will still operate normally. The
  workaround is to rotate the device by 90° twice.

Fixes #18

[dangerous permission]: https://developer.android.com/guide/topics/permissions/requesting.html#perm-groups